gen-ai/examples/evaluations/evaluate_helper_function/pyproject.toml
[project]
name = "evaluate-helper-function"
version = "0.0.0"
description = "Using the evaluate() convenience helper function"
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.0,<2.0.0",
    "gllm-evals[deepeval,langchain,ragas]>=0.0.40",
]

[[tool.uv.index]]
name = "gen-ai-internal"
url = "https://glsdk.gdplabs.id/gen-ai-internal/simple/"

[tool.uv.sources]
gllm-evals = { index = "gen-ai-internal" }
